commit 2

+++ Assets/Scripts/EnemyHealth.cs
@@ -1,11 +1,10 @@
-using System.Collections;
-using System.Collections.Generic;
+
 using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour, IDamagable
 {
-    [SerializeField] int maxHealth;
-    public int currentHealth;
+    [SerializeField] float maxHealth;
+    public float currentHealth;
 
     // Start is called before the first frame update
     void Start()
@@ -22,7 +21,7 @@ public class EnemyHealth : MonoBehaviour, IDamagable
         }
     }
 
-    public void TakeDamage(int damageAmount)
+    public void TakeDamage(float damageAmount)
     {
         currentHealth -= damageAmount;
     }
